refactor(server): move instance_lock_in and the route audit to their callers
`instance_lock_in` is a get-or-create over `AppState::instance_locks`, so
it belongs beside that map rather than in the ACP event listener, where
it forced `state.rs` to import from `acp_events.rs`.

`router_mutating_routes` has one caller, the CityHall exhaustiveness
audit in `access.rs`, so it moves into that test module. `test_helpers`
keeps only the fixture two submodules share.
